// selt-assesment-charts/src/infraestructure/LoadFile/InformaDBAssesment/types.d.ts
export type InformaDBSchemaData = {
  id: number,
  equipo: string,
  multidisciplinar: strint,
  trabajoDesdeBacklog: string,
  mejoraContinua: string,
  coResponsabilidad: string,
  feedbackUsuarios: string,
  coordinacionInterEquipos: string,
  progresoPorTrabajoTerminado: string,
  eventosScrumKanban: string,
  rolesScrumKanban: string,
  backlogGestionadoPriorizadoActualizado: string,
  areasSolicitantes: string,
  canalSolicitud: string,
  comunicacionOtrasAreas: number,
  otrasMejoras: string,
}